Fix a bug in VulkanArena.
We were calling allocator_->free() on an AllocationToken, instead
of deleting it properly. This meant that the iterator
was not getting deleted it was just getting zeroed.

In windows, debug, checked iterators are a lot more than just a
POD type, deleting a checked iterator can modify other iterators
to the same object. (They are internally a big linked list
of all iterators to the same object).