A ton of tiny fixes for tests.
Had to make some of the test names smaller because of
path-length issues on windows.

Has to split any tests that created more than one application.
GAPII was getting unloaded in the middle. This is because
on Windows, we can't ld_preload into the process, so when
the library is closed, gapii gets dumped.