Passthrough application and glsl vs hlsl test flow
The build system has been updated to detect applications that contain
hlsl shaders.

For such applications, DXC and glslc are used to compile those shaders
into SPIR-V. If either of these compilers do not exist, no action will
be taken.

When glslc is used, the "-mfmt=c" option is used to get a C-style array
of 32-bit words that is later used in the pipelines.

A python script is added (spirv_c_mfmt.py) to perform the same task for
the SPIR-V binary output of DXC (note: DXC currently only emits binary
SPIR-V).

The HLSL testing framework is added (written in python). This framework
discovers any applications that include HLSL shaders, run the
application with two different sets of shaders, and compare the pixels
of a specific frame in their output. If the pixels match, the test is
considered a PASS.

Comparison can be done for:
GLSL shaders compiled using glslc
HLSL shaders compiled using glslc
HLSL shaders compiled using DXC