Fix for building Android applications on Windows.
Android Studio deeply nests its external builds. This results
in file paths that are way too long for CMake.

This happens even if you do your build directly in the root directory
of a drive. This is fixed by using numbered directories for everything
when building for Android in an attempt to minimize the file path
as much as possible.