Fix some bugs revealed by recent tests
* Fix vkGetImageSubresourceLayout

Create 2D image with multiple layers or mip levels only when the
physical device format properties say it is allowed. And alway query the
subresource layout for the layer and level with the largest number.

Check subresource layout array pitch only when the image has multiple
layers.

Depth pitch checker is kept around, but not used unless the image type
(const) is changed to 3D. But for now, we don't see any device supports
3D linear image.

* Always enable KHR_swapchain extension, so that swapchain functions can be resolved

* Use a larger timeout for Fence_test

* Fix Event_test, use correct dst layout in the image memory barrier