Fix for MacOS.
Turns out that on MacOS (at least newer machines) we cannot
create a single buffer to rule them all if it needs to
be in coherently mapped memory. So limit the things that
a coherent buffer can do to the things that we are actaully
doing with them for now.