Fix synchronization in async_compute sample
The "mailbox" caused images to be released by the async compute queue
without a corresponding acquire on the render queue, which is illegal.

To fix this issue, whenever the async compute thread sends a buffer to
the mailbox, the existing buffer(s) stay in the mailbox; only the most
recent buffer in the mailbox will actually be rendered, and the
remaining buffers will simply be acquired on the render queue and
immediately released back to the async compute queue.